+++ core/src/test/java/io/cucumber/core/io/ResourceLoaderTest.java
@@ -11,31 +11,31 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 
-public class ResourceLoaderTest {
+class ResourceLoaderTest {
 
     @Test
-    public void loads_resources_from_filesystem_dir() {
+    void loads_resources_from_filesystem_dir() {
         URI uri = FeaturePath.parse("src/test/resources/io/cucumber/core");
         Iterable<Resource> files = new FileResourceLoader().resources(uri, ".properties");
         assertThat(toList(files).size(), is(equalTo(3)));
     }
 
     @Test
-    public void loads_resource_from_filesystem_file() {
+    void loads_resource_from_filesystem_file() {
         URI uri = FeaturePath.parse("src/test/resources/io/cucumber/core/bar.properties");
         Iterable<Resource> files = new FileResourceLoader().resources(uri, ".doesntmatter");
         assertThat(toList(files).size(), is(equalTo(1)));
     }
 
     @Test
-    public void loads_resources_from_jar_on_classpath() {
+    void loads_resources_from_jar_on_classpath() {
         URI uri = FeaturePath.parse("classpath:io/cucumber");
         Iterable<Resource> files = new ClasspathResourceLoader(Thread.currentThread().getContextClassLoader()).resources(uri, ".properties");
         assertThat(toList(files).size(), is(equalTo(4)));
     }
 
     private <T> List<T> toList(Iterable<T> it) {
-        List<T> result = new ArrayList<T>();
+        List<T> result = new ArrayList<>();
         for (T t : it) {
             result.add(t);
         }