+++ core/src/test/java/io/cucumber/core/runtime/InvokerTest.java
@@ -12,22 +12,17 @@ import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class InvokerTest {
+class InvokerTest {
 
     @Test
-    public void doesnt_time_out_if_it_doesnt_take_too_long() throws Throwable {
+    void doesnt_time_out_if_it_doesnt_take_too_long() throws Throwable {
         final Slow slow = new Slow();
-        String what = Invoker.timeout(new Invoker.Callback<String>() {
-            @Override
-            public String call() throws Throwable {
-                return slow.slow(10);
-            }
-        }, 50);
+        String what = Invoker.timeout(() -> slow.slow(10), 50);
         assertThat(what, is(equalTo("slept 10ms")));
     }
 
     @Test
-    public void times_out_if_it_takes_too_long() {
+    void times_out_if_it_takes_too_long() {
         final Slow slow = new Slow();
         Executable testMethod = () -> Invoker.timeout(() -> slow.slow(100), 50);
         TimeoutException expectedThrown = assertThrows(TimeoutException.class, testMethod);
@@ -35,7 +30,7 @@ public class InvokerTest {
     }
 
     @Test
-    public void times_out_infinite_loop_if_it_takes_too_long() {
+    void times_out_infinite_loop_if_it_takes_too_long() {
         final Slow slow = new Slow();
         Executable testMethod = () -> Invoker.timeout((Invoker.Callback<Void>) () -> {
             slow.infinite();
@@ -46,7 +41,7 @@ public class InvokerTest {
     }
 
     @Test
-    public void times_out_infinite_latch_wait_if_it_takes_too_long() {
+    void times_out_infinite_latch_wait_if_it_takes_too_long() {
         final Slow slow = new Slow();
         Executable testMethod = () -> Invoker.timeout((Invoker.Callback<Void>) () -> {
             slow.infiniteLatchWait();
@@ -58,30 +53,25 @@ public class InvokerTest {
 
 
     @Test
-    public void times_out_busy_wait_if_it_takes_too_long() {
+    void times_out_busy_wait_if_it_takes_too_long() {
         final Slow slow = new Slow();
-        Executable testMethod = () -> Invoker.timeout((Invoker.Callback<Void>) () -> {
+        Executable testMethod = () -> Invoker.timeout(() -> {
             slow.busyWait();
-            return null;
+            return "foo";
         }, 1);
         TimeoutException expectedThrown = assertThrows(TimeoutException.class, testMethod);
         assertThat(expectedThrown.getMessage(), is(equalTo("Timed out after 1ms.")));
     }
 
     @Test
-    public void doesnt_leak_threads() throws Throwable {
+    void doesnt_leak_threads() throws Throwable {
 
         long initialNumberOfThreads = Thread.getAllStackTraces().size();
         long currentNumberOfThreads = Long.MAX_VALUE;
 
         boolean cleanedUp = false;
         for (int i = 0; i < 1000; i++) {
-            Invoker.timeout(new Invoker.Callback<String>() {
-                @Override
-                public String call() throws Throwable {
-                    return null;
-                }
-            }, 10);
+            Invoker.timeout(() -> "foo", 10);
             Thread.sleep(5);
             currentNumberOfThreads = Thread.getAllStackTraces().size();
             if (i > 20 && currentNumberOfThreads <= initialNumberOfThreads) {
@@ -94,25 +84,26 @@ public class InvokerTest {
             cleanedUp, is(equalTo(true)));
     }
 
-    public static class Slow {
+    static class Slow {
         int busyCounter = Integer.MIN_VALUE;
 
-        public String slow(int millis) throws InterruptedException {
+        String slow(int millis) throws InterruptedException {
             sleep(millis);
             return String.format("slept %sms", millis);
         }
 
-        public void infinite() throws InterruptedException {
+        @SuppressWarnings("InfiniteLoopStatement")
+        void infinite() throws InterruptedException {
             while (true) {
                 sleep(1);
             }
         }
 
-        public void infiniteLatchWait() throws InterruptedException {
+        void infiniteLatchWait() throws InterruptedException {
             new CountDownLatch(1).await();
         }
 
-        public int busyWait() throws InterruptedException {
+        int busyWait() {
             while (busyCounter < Integer.MAX_VALUE) {
                 busyCounter += 1;
             }