+++ examples/java-wicket/java-wicket-main/src/main/java/io/cucumber/examples/java/wicket/view/Rent.java
@@ -19,7 +19,7 @@ public class Rent extends WebPage {
 
     private class RentCarForm extends Form<ValueMap> {
         public RentCarForm(String id) {
-            super(id, new CompoundPropertyModel<ValueMap>(new ValueMap()));
+            super(id, new CompoundPropertyModel<>(new ValueMap()));
         }
 
         @Override