+++ java8/src/test/java/io/cucumber/java8/Java8AnonInnerClassStepDefinitionTest.java
@@ -8,16 +8,17 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 
-public class Java8AnonInnerClassStepDefinitionTest {
+@SuppressWarnings("Convert2Lambda")
+class Java8AnonInnerClassStepDefinitionTest {
 
     @Test
-    public void should_calculate_parameters_count_from_body_with_one_param() {
+    void should_calculate_parameters_count_from_body_with_one_param() {
         Java8StepDefinition java8StepDefinition = Java8StepDefinition.create("I have some step", StepdefBody.A1.class, oneParamStep());
         assertThat(java8StepDefinition.parameterInfos().size(), is(equalTo(1)));
     }
 
     @Test
-    public void should_calculate_parameters_count_from_body_with_two_params() {
+    void should_calculate_parameters_count_from_body_with_two_params() {
         Java8StepDefinition java8StepDefinition = Java8StepDefinition.create("I have some step", StepdefBody.A2.class, twoParamStep());
         assertThat(java8StepDefinition.parameterInfos().size(), is(equalTo(2)));
     }